Move HappBundle into conductor_lib
